Optimise scripts under folder `ip`
