+++ tests/ntp/test_ntp.py
@@ -1,9 +1,10 @@
 from tests.common.utilities import wait_until
 from tests.common.helpers.assertions import pytest_assert
+from tests.common.helpers.ntp_helper import check_ntp_status, run_ntp, _context_for_setup_ntp
 import logging
 import time
 import pytest
-from contextlib import contextmanager
+
 
 logger = logging.getLogger(__name__)
 
@@ -53,49 +54,6 @@ def setup_ntp(ptfhost, duthosts, rand_one_dut_hostname, ptf_use_ipv6):
         yield result
 
 
-@pytest.fixture(scope="function")
-def setup_ntp_func(ptfhost, duthosts, rand_one_dut_hostname, ptf_use_ipv6):
-    with _context_for_setup_ntp(ptfhost, duthosts, rand_one_dut_hostname, ptf_use_ipv6) as result:
-        yield result
-
-
-@contextmanager
-def _context_for_setup_ntp(ptfhost, duthosts, rand_one_dut_hostname, ptf_use_ipv6):
-    """setup ntp client and server"""
-    duthost = duthosts[rand_one_dut_hostname]
-
-    ptfhost.lineinfile(path="/etc/ntp.conf", line="server 127.127.1.0 prefer")
-
-    # restart ntp server
-    ntp_en_res = ptfhost.service(name="ntp", state="restarted")
-
-    pytest_assert(wait_until(120, 5, 0, check_ntp_status, ptfhost),
-                  "NTP server was not started in PTF container {}; NTP service start result {}"
-                  .format(ptfhost.hostname, ntp_en_res))
-
-    # setup ntp on dut to sync with ntp server
-    config_facts = duthost.config_facts(host=duthost.hostname, source="running")['ansible_facts']
-    ntp_servers = config_facts.get('NTP_SERVER', {})
-    for ntp_server in ntp_servers:
-        duthost.command("config ntp del %s" % ntp_server)
-
-    duthost.command("config ntp add %s" % (ptfhost.mgmt_ipv6 if ptf_use_ipv6 else ptfhost.mgmt_ip))
-
-    yield
-
-    # stop ntp server
-    ptfhost.service(name="ntp", state="stopped")
-    # reset ntp client configuration
-    duthost.command("config ntp del %s" % (ptfhost.mgmt_ipv6 if ptf_use_ipv6 else ptfhost.mgmt_ip))
-    for ntp_server in ntp_servers:
-        duthost.command("config ntp add %s" % ntp_server)
-    # The time jump leads to exception in lldp_syncd. The exception has been handled by lldp_syncd,
-    # but it will leave error messages in syslog, which will cause subsequent test cases to fail.
-    # So we need to wait for a while to make sure the error messages are flushed.
-    # The default update interval of lldp_syncd is 10 seconds, so we wait for 20 seconds here.
-    time.sleep(20)
-
-
 @pytest.fixture
 def setup_long_jump_config(duthosts, rand_one_dut_hostname):
     """set long jump config and set DUT's time forward"""
@@ -127,13 +85,6 @@ def setup_long_jump_config(duthosts, rand_one_dut_hostname):
     config_long_jump(duthost, long_jump_enable)
 
 
-def check_ntp_status(host):
-    res = host.command("ntpstat", module_ignore_errors=True)
-    if res['rc'] != 0:
-        return False
-    return True
-
-
 def test_ntp_long_jump_enabled(duthosts, rand_one_dut_hostname, setup_ntp, setup_long_jump_config):
     duthost = duthosts[rand_one_dut_hostname]
 
@@ -150,24 +101,6 @@ def test_ntp_long_jump_disabled(duthosts, rand_one_dut_hostname, setup_ntp, setu
 
     pytest_assert(wait_until(720, 10, 0, check_ntp_status, duthost),
                   "NTP long jump disable failed")
-
-
-def run_ntp(duthosts, rand_one_dut_hostname, setup_ntp):
-    """ Verify that DUT is synchronized with configured NTP server """
-    duthost = duthosts[rand_one_dut_hostname]
-
-    ntpsec_conf_stat = duthost.stat(path="/etc/ntpsec/ntp.conf")
-    using_ntpsec = ntpsec_conf_stat["stat"]["exists"]
-
-    duthost.service(name='ntp', state='stopped')
-    if using_ntpsec:
-        duthost.command("timeout 20 ntpd -gq -u ntpsec:ntpsec")
-    else:
-        ntp_uid = ":".join(duthost.command("getent passwd ntp")['stdout'].split(':')[2:4])
-        duthost.command("timeout 20 ntpd -gq -u {}".format(ntp_uid))
-    duthost.service(name='ntp', state='restarted')
-    pytest_assert(wait_until(720, 10, 0, check_ntp_status, duthost),
-                  "NTP not in sync")
 
 
 def test_ntp(duthosts, rand_one_dut_hostname, setup_ntp):
