+++ tests/tacacs/test_authorization.py
@@ -4,9 +4,9 @@ import pytest
 from _pytest.outcomes import Failed
 import time
 
-from tests.tacacs.utils import stop_tacacs_server, start_tacacs_server
-from tests.tacacs.utils import per_command_authorization_skip_versions, \
-        remove_all_tacacs_server, get_ld_path, change_and_wait_aaa_config_update, \
+from tests.common.helpers.tacacs_helper import stop_tacacs_server, start_tacacs_server, \
+    per_command_authorization_skip_versions, remove_all_tacacs_server, get_ld_path
+from tests.tacacs.utils import change_and_wait_aaa_config_update, \
         ensure_tacacs_server_running_after_ut                             # noqa: F401
 from tests.common.helpers.assertions import pytest_assert
 from tests.common.utilities import skip_release, wait_until, paramiko_ssh
